Extract utils as a separate crate
