+++ ecmascript/transforms/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "swc_ecma_transforms"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["강동윤 <kdy1997.dev@gmail.com>"]
 license = "Apache-2.0/MIT"
 repository = "https://github.com/swc-project/swc.git"
@@ -12,6 +12,7 @@ edition = "2018"
 swc_atoms = { version = "0.2.0", path ="../../atoms" }
 swc_common = { version = "0.4.2", path ="../../common" }
 ast = { package = "swc_ecma_ast", version = "0.14.0", path ="../ast" }
+utils = { package = "swc_ecma_utils", version = "0.1.0", path ="../utils" }
 swc_ecma_parser = { version = "0.16", path ="../parser", features = ["verify"] }
 chashmap = "2.2.0"
 either = "1.5"
