+++ ecmascript/utils/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "swc_ecma_utils"
+version = "0.1.0"
+authors = ["강동윤 <kdy1997.dev@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ast = { package = "swc_ecma_ast", version = "0.14.0", path ="../ast" }
+swc_atoms = { version = "0.2.0", path ="../../atoms" }
+swc_common = { version = "0.4.2", path ="../../common" }
+swc_ecma_parser = { version = "0.16", path ="../parser", features = ["verify"] }
+lazy_static = "1"
+scoped-tls = "1"
+unicode-xid = "0.2"
